Collision Component + Bug fixes
The collision components allow the user to  add collision to objects. The default collision is a box. 
The bugs that have been fixed are, the selection boxes (enums) previously changed the selection on all entities (just what can be seen in editor) and when reading in a scene file the boxes wouldn't be the right selection.
The other bug is the PhysX visual debugger did not get reset everytime the game gets played and stopped